pundit for creating new rents

+++ app/controllers/api/v1/rents_controller.rb
@@ -2,15 +2,17 @@ module Api
   module V1
     class RentsController < ApplicationController
       def index
-        render json: current_user.rents.page(params[:page])
+        render json: policy_scope(Rent).page(params[:page])
       end
 
       def show
+        authorize rent
         render json: rent
       end
 
       def create
-        @rent = current_user.rents.build(rent_params)
+        @rent = Rent.new(rent_params)
+        authorize rent
         if rent.save
           head :created
         else
@@ -19,6 +21,7 @@ module Api
       end
 
       def destroy
+        authorize rent
         rent.destroy
         head :ok
       end
@@ -26,7 +29,7 @@ module Api
       private
 
       def rent_params
-        params.require(:rent).permit(:from, :to, :book_id)
+        params.require(:rent).permit(:from, :to, :book_id, :user_id)
       end
 
       def rent
